add PH sensor and soil moisture sensor calibration

+++ esp32/sensorsManager.h
@@ -33,7 +33,7 @@ class SensorsManager {
         float getHumidity();
 
         // Soil moisture sensor
-        int getSoilMoisture();
+        float getSoilMoisture();
 
         // PH sensor
         float getPH();
